refactor(workout-plan): extract execution-style picker to workout-plan-execution-style.js (WP3.4d)
Mechanical feature-split from the workout-plan.js monolith: move the
execution-style options cache and the AMRAP/EMOM/standard picker overlay
into a new static/js/modules/workout-plan-execution-style.js. Move-only —
no behavior, API, DB, calculation, response-shape, or event-timing change.

Moved verbatim (bodies byte-identical to HEAD, AST/diff-verified):
- executionStyleOptions module state
- getExecutionStyleOptions (internal cache)
- showExecutionStylePicker (exported)

Two intended edits only:
- showExecutionStylePicker gained the `export` keyword so the monolith can
  keep injecting it into the table via configureWorkoutPlanTable({
  onExecutionStyleClick }).
- Its post-save fetchWorkoutPlan() (orchestration that stays in the monolith)
  is resolved by callback injection — configureExecutionStyle({ refreshPlan })
  called once at monolith load, default no-op — mirroring WP3.4b's table
  split, so the new module never imports back into workout-plan.js (no
  circular import).

buildExecutionStylePayload is reused from workout-plan-helpers.js and pruned
from the monolith's now-unused import.

Gates: Vitest 88/88; E2E workout-plan + exercise-interactions +
superset-edge-cases + fatigue-context + replace-exercise-errors 77 passed;
learned-calibration 8/8 isolated.
